refactor(input-selection)!: renamed cip2 package to input-selection
