+++ packages/cardano-services/test/tsconfig.json
@@ -5,7 +5,7 @@
   },
   "include": [
     "./**/*.ts",
-    "../../cip2/test/util",
+    "../../input-selection/test/util",
     "../../core/test/CardanoNode",
     "../../ogmios/test/mocks"
   ],
@@ -29,7 +29,7 @@
       "path": "../../ogmios/src"
     },
     {
-      "path": "../../cip2/src"
+      "path": "../../input-selection/src"
     }
   ]
 }
