+++ packages/input-selection/LICENSE
