+++ packages/input-selection/src/InputSelectionError.ts
