+++ packages/input-selection/src/selectionConstraints.ts
