+++ packages/input-selection/src/tsconfig.json
