+++ packages/input-selection/test/selectionConstraints.test.ts
