+++ packages/input-selection/test/util/properties.ts
