+++ packages/wallet/package.json
@@ -72,9 +72,9 @@
     "webextension-polyfill": "^0.9.0"
   },
   "dependencies": {
-    "@cardano-sdk/cip2": "^0.5.0",
     "@cardano-sdk/cip30": "^0.5.0",
     "@cardano-sdk/core": "^0.5.0",
+    "@cardano-sdk/input-selection": "^0.5.0",
     "@cardano-sdk/key-management": "^0.1.0",
     "@cardano-sdk/util": "^0.5.0",
     "@cardano-sdk/util-rxjs": "^0.4.1",
